# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_shares.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  internal_shares.ShareOverview:
    properties:
      config: {}
      name:
        type: string
      path:
        type: string
      type:
        description: '"smb" or "nfs"'
        type: string
    type: object
...
  internal_users.Role:
    enum:
    - admin
    - user
    - guest
    type: string
    x-enum-comments:
      RoleAdmin: 管理员：全部权限
      RoleGuest: 访客：只读访问
      RoleUser: 普通用户：受限访问
    x-enum-descriptions:
    - 管理员：全部权限
    - 普通用户：受限访问
    - 访客：只读访问
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleGuest
  internal_users.User:
    properties:
      created_at:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      groups:
        description: 所属用户组
        items:
          type: string
        type: array
      home_dir:
        description: 用户主目录
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/internal_users.Role'
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_users.UserInput:
    properties:
      email:
        type: string
      groups:
        items:
          type: string
        type: array
      home_dir:
        type: string
      password:
        minLength: 6
        type: string
      role:
        $ref: '#/definitions/internal_users.Role'
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  internal_web.GenericResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
...
  internal_web.VolumeCreateRequest:
    properties:
      devices:
        example:
        - /dev/sda
        - /dev/sdb
        items:
          type: string
        type: array
      name:
        example: data-vol
        type: string
      profile:
        description: single, raid0, raid1, raid10, raid5, raid6
        example: raid1
        type: string
    required:
    - devices
    - name
    type: object
  nas-os_internal_api.Response:
    properties:
      code:
        description: 业务码：0=成功，非0=失败
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
  nas-os_internal_nfs.Client:
    properties:
      host:
        type: string
      options:
        items:
          type: string
        type: array
    type: object
  nas-os_internal_nfs.Export:
    properties:
      clients:
        items:
          $ref: '#/definitions/nas-os_internal_nfs.Client'
        type: array
      comment:
        type: string
      fsid:
        type: integer
      options:
        $ref: '#/definitions/nas-os_internal_nfs.ExportOptions'
      path:
        type: string
    type: object
  nas-os_internal_nfs.ExportOptions:
    properties:
      async:
        type: boolean
      no_root_squash:
        type: boolean
      ro:
        type: boolean
      rw:
        type: boolean
      secure:
        type: boolean
      subtree_check:
        type: boolean
    type: object
  nas-os_internal_nfs.ExportRequest:
    properties:
      clients:
        items:
          $ref: '#/definitions/nas-os_internal_nfs.Client'
        type: array
      comment:
        type: string
      fsid:
        type: integer
      options:
        properties:
          async:
            type: boolean
          no_root_squash:
            type: boolean
          ro:
            type: boolean
          rw:
            type: boolean
          secure:
            type: boolean
          subtree_check:
            type: boolean
        type: object
      path:
        type: string
    required:
    - path
    type: object
  nas-os_internal_smb.Share:
    properties:
      acl_group_control:
        type: boolean
      allowed_users:
        items:
          type: string
        type: array
      available:
        type: boolean
      browseable:
        type: boolean
      comment:
        type: string
      create_mask:
        type: string
      directory_mask:
        type: string
      ea_support:
        type: boolean
      follow_symlinks:
        type: boolean
      force_group:
        type: string
      force_user:
        type: string
      guest_access:
        type: boolean
      guest_ok:
        type: boolean
      hide_files:
        items:
          type: string
        type: array
      inherit_acls:
        type: boolean
      inherit_owner:
        type: string
      inherit_permissions:
        type: boolean
      name:
        type: string
      path:
        type: string
      printable:
        type: boolean
      read_only:
        type: boolean
      store_dos_attributes:
        type: boolean
      users:
        items:
          type: string
        type: array
      valid_users:
        items:
          type: string
        type: array
      veto_files:
        items:
          type: string
        type: array
      wide_links:
        type: boolean
      write_list:
        items:
          type: string
        type: array
    type: object
  nas-os_internal_smb.ShareInput:
    properties:
      allowed_users:
        items:
          type: string
        type: array
      browseable:
        type: boolean
      comment:
        type: string
      create_mask:
        type: string
      directory_mask:
        type: string
      guest_access:
        type: boolean
      guest_ok:
        type: boolean
      name:
        type: string
      path:
        type: string
      read_only:
        type: boolean
      users:
        items:
          type: string
        type: array
      valid_users:
        items:
          type: string
        type: array
      veto_files:
        items:
          type: string
        type: array
      write_list:
        items:
          type: string
        type: array
    required:
    - name
    - path
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nas-os.io
    name: NAS-OS 团队
    url: https://github.com/nas-os/nas-os
  description: |-
    NAS-OS 是一个现代化的网络存储操作系统，提供卷管理、用户管理、共享管理、网络配置等功能。

    ## 功能模块
    - **卷管理**: Btrfs 卷创建、快照、RAID 配置
    - **用户管理**: 用户/组管理、认证授权
    - **共享管理**: SMB/NFS 共享配置
    - **网络管理**: 网络接口、DDNS、防火墙、端口转发
    - **Docker 管理**: 容器、镜像、应用商店
    - **插件系统**: 可扩展的插件架构
    - **配额管理**: 存储配额控制
    - **性能监控**: 系统性能监控与报告
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NAS-OS API
...
paths:
...
  /shares:
    get:
      consumes:
      - application/json
      description: 获取所有 SMB 和 NFS 共享的统一列表
      produces:
      - application/json
      responses:
        "200":
          description: 共享列表
          schema:
            allOf:
            - $ref: '#/definitions/internal_shares.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_shares.ShareOverview'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_shares.Response'
      security:
      - BearerAuth: []
      summary: 列出所有共享
      tags:
      - shares
  /shares/nfs:
    get:
      consumes:
      - application/json
      description: 获取所有 NFS 导出列表
      produces:
      - application/json
      responses:
        "200":
          description: NFS 导出列表
          schema:
            allOf:
            - $ref: '#/definitions/internal_shares.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/nas-os_internal_nfs.Export'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
      security:
      - BearerAuth: []
      summary: 列出 NFS 导出
      tags:
      - shares
    post:
      consumes:
      - application/json
      description: 创建新的 NFS 导出
      parameters:
      - description: 导出配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/nas-os_internal_nfs.ExportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_shares.Response'
            - properties:
                data:
                  $ref: '#/definitions/nas-os_internal_nfs.Export'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "409":
          description: 导出已存在
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
      security:
      - BearerAuth: []
      summary: 创建 NFS 导出
      tags:
      - shares
  /shares/smb:
    get:
      consumes:
      - application/json
      description: 获取所有 SMB 共享列表
      produces:
      - application/json
      responses:
        "200":
          description: SMB 共享列表
          schema:
            allOf:
            - $ref: '#/definitions/internal_shares.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/nas-os_internal_smb.Share'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
      security:
      - BearerAuth: []
      summary: 列出 SMB 共享
      tags:
      - shares
    post:
      consumes:
      - application/json
      description: 创建新的 SMB 共享
      parameters:
      - description: 共享配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/nas-os_internal_smb.ShareInput'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_shares.Response'
            - properties:
                data:
                  $ref: '#/definitions/nas-os_internal_smb.Share'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "409":
          description: 共享已存在
          schema:
            $ref: '#/definitions/internal_shares.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_shares.Response'
      security:
      - BearerAuth: []
      summary: 创建 SMB 共享
      tags:
      - shares
...
  /users:
    get:
      consumes:
      - application/json
      description: 获取系统中所有用户列表，支持按角色筛选
      parameters:
      - description: 按角色筛选 (admin, user)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            allOf:
            - $ref: '#/definitions/nas-os_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_users.User'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
      security:
      - BearerAuth: []
      summary: 列出所有用户
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建新的系统用户，需要管理员权限
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_users.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/nas-os_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_users.User'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/nas-os_internal_api.Response'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - users
...
  /volumes:
    get:
      consumes:
      - application/json
      description: 获取系统中所有 Btrfs 卷的列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 列出所有卷
      tags:
      - volumes
    post:
      consumes:
      - application/json
      description: 使用指定设备和配置创建新的 Btrfs 卷
      parameters:
      - description: 卷创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_web.VolumeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 创建新卷
      tags:
      - volumes
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API 密钥认证
    in: query
    name: token
    type: apiKey
  BearerAuth:
    description: JWT 认证令牌，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...